backend/requirements.txt
fastapi==0.115.0
uvicorn==0.30.0
httpx==0.27.0
pydantic==2.9.0
pydantic-settings==2.5.2
PyJWT[crypto]>=2.8.0
python-dotenv>=1.0.0
